refactor align.beam
