+++ mfx/beamline.py
@@ -82,9 +82,9 @@ with safe_load('OM'):
     from mfx.om import *
     om = OM()
 
-with safe_load('align'):
-    from mfx.optimize.align import *
-    align = Align()
+with safe_load('beam'):
+    from mfx.optimize.beam import *
+    beam = Beam()
 
 with safe_load('vernier'):
     from mfx.vernier import *
